+++ scripts/intel_memory_cache_model.py
@@ -20,12 +20,12 @@ def parse_and_validate_args() -> argparse.Namespace:
     parser.add_argument(
         "--kit",
         type=str,
-        default=".",
+        default="./GTPin",
         help="Full path of the GTPin kit",
     )
     parser.add_argument(
         "--app", type=str,
-        default=".",
+        default="./pti-gpu/samples/ze_gemm/build",
         help="Full path of the compiled ze_gemm application",
     )
 
@@ -74,9 +74,14 @@ def get_hits_misses_count_tool_results(
 
 def main():
     args           = parse_and_validate_args()
-    gtpin          = os.path.join(args.kit, "Profilers", "Bin", "gtpin")
-    ze_gemm_folder = os.path.join(args.app)
+    gtpin          = os.path.realpath(os.path.join(args.kit, "Profilers", "Bin", "gtpin"))
+    ze_gemm_folder = os.path.realpath(os.path.join(args.app))
     result_path    = os.path.join(ze_gemm_folder, RESULT_PKLE_FILE_NAME)
+
+    if not os.path.isfile(gtpin):
+        raise FileNotFoundError(
+            f"Could not find GTPin profiler at {gtpin}. Please check the path and try again."
+        )
 
     pickl_file = open(result_path, "wb")
 
@@ -105,7 +110,7 @@ def main():
                 # ------------------------------------------------------------------
                 print("Running: {}".format(' '.join(cmd)))
 
-                rc, stdout, stderr = capture_subprocess_output(cmd=cmd, cwd=ze_gemm_folder)
+                rc, stdout, stderr = capture_subprocess_output(cmd=cmd)
                 if rc:
                     raise ChildProcessError("Profiling run failed. rc = {}".format(rc))
